move withSettings HoC into its module
