+++ src/frontend/web_application/src/modules/settings/hoc/withSettings.js
@@ -1,6 +1,6 @@
 import { createSelector } from 'reselect';
 import { connect } from 'react-redux';
-import { settingsSelector } from '../store/selectors/settings';
+import { settingsSelector } from '../../../store/selectors/settings';
 
 const mapStateToProps = createSelector(
   [settingsSelector],
@@ -11,4 +11,4 @@ const mapStateToProps = createSelector(
 
 const withSettings = () => Component => connect(mapStateToProps)(Component);
 
-export { withSettings };
+export default withSettings;
