+++ src/frontend/web_application/src/modules/settings/index.js
@@ -1,1 +1,2 @@
 export { default as WithSettings } from './components/WithSettings';
+export { default as withSettings } from './hoc/withSettings';
