+++ src/frontend/web_application/src/scenes/Signup/index.js
@@ -2,7 +2,7 @@ import { bindActionCreators, compose } from 'redux';
 import { connect } from 'react-redux';
 import { withTranslator } from '@gandi/react-translate';
 import { push } from 'react-router-redux';
-import { withSettings } from '../../hoc/settings';
+import { withSettings } from '../../modules/settings';
 import Presenter from './presenter';
 
 const mapDispatchToProps = dispatch => bindActionCreators({ onSignupSuccess: push }, dispatch);
