+++ src/frontend/web_application/src/scenes/Timeline/components/MessageItem/index.js
@@ -1,6 +1,6 @@
 import { compose } from 'redux';
 import { withTranslator } from '@gandi/react-translate';
-import { withSettings } from '../../../../hoc/settings';
+import { withSettings } from '../../../../modules/settings';
 import Presenter from './presenter';
 
 export default compose(
